Don't delete file contents on error
When running `ansible::{en,de}crypt-buffer` if the vault password is
wrong, an error is raised and the contents of the buffer are deleted.

This is annoying to me, since I have to deal with various files with
different vault passwords.

This PR fixes this behavior by trying to decrypt the buffer *before*
deleting its contents.