Add `ansible::auto-decrypt-encrypt`
`ansible::auto-decrypt-encrypt` function is meant to be added to
`ansible-hook`.

This function will try to decrypt the visited file. If the file was
successfully decrypted, `ansible::encrypt-buffer` will be added to the
local `before-save-hook`, and `ansible::decrypt-buffer` will be added to
the local `after-save-hook`. That means that the visited file will be
encrypted just before being saved to disk, and will be decrypted right
afterwards, leading to an experience close to how easypg works.

Closes #9