refactor!: rename keybinding macros to be consistent with macro names
In MinEmacs, I'm trying to follow a convention: every function/variable
should either start with `minemacs-` giving it a clear name space, or
start with `+` to tell it is an added function from MinEmacs.

For macros, I've adopted the convention I've seen before in Doom Emacs,
which is the use of the `!` postfix to indicate that this is a macro and
not a function.

This commit renames the keybinding macros to follow this convention. It
WILL BREAK THE CURRENT USER CONFIGURATIONS.