fix(magit-file-icons): load before `magit`, but not too early
This also fix an issue related to the annoying `track-changes` and
`parinfer-rust-mode`. For some reason, when opening Magit from an Elisp buffer,
it enables the `emacs-lisp-mode` on the `magit-status` buffer. This causes
`parinfer-rust-mode` (and `track-changes`) to be setup on the buffer. After
doing some changes in the original Elisp buffer, this triggers
`track-changes--tracker-signal` in the `magit-status` buffer, and the latter
being read-only will trigger an error.

Loading `magit-file-icons`, which loads `magit` and advises it, seems to fix
this issue.